# openapi.yaml
openapi: 3.1.0
info:
  title: Store API
  version: 1.0.0
  description: |
    Contrato derivado de los controladores, DTO, servicios, seguridad y manejadores
    de excepciones del backend Spring Boot. La API usa JSON y autenticación JWT Bearer.

    La paginación canónica usa `page` con índice inicial 1. Por compatibilidad, también
    se acepta `pageIndex` con índice inicial 0; cuando ambos se envían, `pageIndex`
    prevalece. No existe ordenamiento configurable en los endpoints actuales.
servers:
  - url: http://localhost:8080/api
    description: Entorno local
...
paths:
...
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Listar todos los usuarios
      operationId: listarUsuarios
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuarios registrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [Usuarios]
      summary: Registrar usuario o crear usuario como administrador
      description: |
        Un visitante siempre crea un usuario `CLIENTE`, incluso si envía `role`.
        Un administrador autenticado puede asignar `ADMIN` o `CLIENTE`.
      operationId: crearUsuario
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRequest'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /usuarios/me:
    get:
      tags: [Usuarios]
      summary: Obtener el perfil autenticado
      operationId: obtenerMiPerfil
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Usuarios]
      summary: Actualizar el perfil autenticado
      description: No permite cambiar el rol. La contraseña se reemplaza por el valor enviado.
      operationId: actualizarMiPerfil
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarPerfilRequest'
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
    delete:
      tags: [Usuarios]
      summary: Eliminar la cuenta autenticada
      operationId: eliminarMiCuenta
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Cuenta eliminada
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /usuarios/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Usuarios]
      summary: Obtener usuario por ID
      operationId: obtenerUsuario
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Usuarios]
      summary: Actualizar un usuario
      operationId: actualizarUsuario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRequest'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
    delete:
      tags: [Usuarios]
      summary: Eliminar un usuario
      operationId: eliminarUsuario
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Usuario eliminado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /productos:
    get:
      tags: [Productos]
      summary: Listar productos paginados
      operationId: listarProductos
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Página de productos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaProductoResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
    post:
      tags: [Productos]
      summary: Crear un producto
      operationId: crearProducto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /productos/buscar:
    get:
      tags: [Productos]
      summary: Buscar productos por nombre
      operationId: buscarProductos
      security: []
      parameters:
        - name: nombre
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Productos cuyo nombre contiene el texto, sin distinguir mayúsculas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaProductoResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /productos/filtrar:
    get:
      tags: [Productos]
      summary: Filtrar productos
      operationId: filtrarProductos
      security: []
      parameters:
        - name: nombre
          in: query
          required: false
          schema:
            type: string
        - name: categoriaId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: precioMin
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: precioMax
          in: query
          required: false
          schema:
            type: number
            format: double
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Página de productos filtrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaProductoResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /productos/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Productos]
      summary: Obtener producto por ID
      operationId: obtenerProducto
      security: []
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Productos]
      summary: Reemplazar los datos de un producto
      operationId: actualizarProducto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Productos]
      summary: Eliminar un producto
      operationId: eliminarProducto
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Producto eliminado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /productos/{id}/stock:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [Productos]
      summary: Actualizar únicamente el stock
      operationId: actualizarStockProducto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockRequest'
      responses:
        '200':
          description: Producto con stock actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token emitido por `/auth/login`; expira a las 24 horas y contiene id, email, role y tokenVersion.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      description: Página con índice inicial 1. El valor 0 también representa la primera página.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 1
    PageIndex:
      name: pageIndex
      in: query
      description: Alternativa compatible con índice inicial 0; prevalece sobre `page`.
      required: false
      schema:
        type: integer
        minimum: 0
    Size:
      name: size
      in: query
      description: Cantidad de elementos por página. El backend no impone un máximo.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 10
  responses:
    BadRequestError:
      description: Solicitud inválida o regla de negocio incumplida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Error de validación del cuerpo; también puede ser una regla de negocio inválida
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ValidationErrorResponse'
              - $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Credenciales incorrectas, JWT ausente, inválido, expirado o invalidado por versión
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Usuario autenticado sin el rol requerido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Recurso duplicado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UsuarioRequest:
      type: object
      additionalProperties: false
      required: [nombre, email, password]
      properties:
        nombre:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
        role:
          oneOf:
            - $ref: '#/components/schemas/Role'
            - type: 'null'
          description: Opcional. Solo se respeta cuando quien crea es ADMIN.
    ActualizarPerfilRequest:
      type: object
      additionalProperties: false
      required: [nombre, email, password]
      properties:
        nombre:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          description: Contraseña nueva; el backend no solicita la contraseña anterior.
    UsuarioResponse:
      type: object
      required: [id, nombre, email, role]
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
    Role:
      type: string
      enum: [ADMIN, CLIENTE]
...
    ProductoRequest:
      type: object
      additionalProperties: false
      required: [nombre, precio, stock, categoriaId]
      properties:
        nombre:
          type: string
          minLength: 1
        precio:
          type: number
          format: double
          minimum: 0
        stock:
          type: integer
          minimum: 0
        categoriaId:
          type: integer
          format: int64
    ProductoResponse:
      type: object
      required: [id, nombre, precio, stock, categoriaId, categoriaNombre]
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        precio:
          type: number
          format: double
        stock:
          type: integer
        categoriaId:
          type: integer
          format: int64
        categoriaNombre:
          type: string
    StockRequest:
      type: object
      additionalProperties: false
      required: [stock]
      properties:
        stock:
          type: integer
          minimum: 0
...
    PaginaProductoResponse:
      allOf:
        - $ref: '#/components/schemas/PaginaBase'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ProductoResponse'
...
    PaginaBase:
      type: object
      required: [content, totalElements, totalPages, currentPage, size]
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
    ErrorResponse:
      type: object
      required: [timestamp, status, error, message]
      properties:
        timestamp:
          type: string
          description: Fecha y hora local serializada como texto ISO-8601.
        status:
          type: integer
        error:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          required: [errors]
          properties:
            errors:
              type: object
              additionalProperties:
                type: string

# TODO: El backend no incluye un manejador explícito para errores de conversión de parámetros
# (por ejemplo, `page=abc`) ni para errores internos de persistencia. Spring generará esas
# respuestas con su formato por defecto, que puede variar entre versiones; no se documenta un
# schema estable para esos casos hasta que el backend defina un contrato explícito.