# src/main/java/com/example/store/usuario/UsuarioController.java
package com.example.store.usuario;

import com.example.store.usuario.dto.ActualizarPerfilRequest;
import com.example.store.usuario.dto.UsuarioRequest;
import com.example.store.usuario.dto.UsuarioResponse;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/usuarios")
public class UsuarioController {

    private final UsuarioService usuarioService;

    public UsuarioController(UsuarioService usuarioService) {
        this.usuarioService = usuarioService;
    }

    @GetMapping
    @PreAuthorize("hasRole('ADMIN')")
    public List<UsuarioResponse> listarTodos() {
        return usuarioService.listarTodos();
    }

    @PostMapping
    @PreAuthorize("isAnonymous() or hasRole('ADMIN')")
    @ResponseStatus(HttpStatus.CREATED)
    public UsuarioResponse crear(@Valid @RequestBody UsuarioRequest request) {
        return usuarioService.crear(request);
    }

    @GetMapping("/me")
    public UsuarioResponse obtenerMiPerfil() {
        return usuarioService.obtenerMiPerfil();
    }

    @PutMapping("/me")
    public UsuarioResponse actualizarMiPerfil(@Valid @RequestBody ActualizarPerfilRequest request) {
        return usuarioService.actualizarMiPerfil(request);
    }

    @DeleteMapping("/me")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void eliminarMiCuenta() {
        usuarioService.eliminarMiCuenta();
    }

    @GetMapping("/{id}")
    @PreAuthorize("hasRole('ADMIN')")
    public UsuarioResponse obtener(@PathVariable Long id) {
        return usuarioService.obtenerPorId(id);
    }

    @PutMapping("/{id}")
    @PreAuthorize("hasRole('ADMIN')")
    public UsuarioResponse actualizar(@PathVariable Long id, @Valid @RequestBody UsuarioRequest request) {
        return usuarioService.actualizar(id, request);
    }

    @DeleteMapping("/{id}")
    @PreAuthorize("hasRole('ADMIN')")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void eliminar(@PathVariable Long id) {
        usuarioService.eliminar(id);
    }
}